+++ src/Vector2.hpp
@@ -13,7 +13,7 @@ public:
 	float x;
 	float y;
 
-	Vector2(int x = 0, int y = 0);
+	Vector2(float x = 0, float y = 0);
 
 	float abs() const;
